+++ src/native/components/Transactions/Edit.tsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import {
+  ScrollView,
+  KeyboardAvoidingView,
+} from 'native-base';
+
+import Title from '../UI/Title';
+import TransactionForm from './Form';
+
+const Edit = ({
+  loading,
+  payload,
+  accounts,
+  submit,
+  getAutocompleteAccounts,
+  loadingAutocomplete,
+  navigation,
+  goToTransactions,
+}) => (
+  <>
+    <Title navigation={navigation} text={payload.description} />
+    <KeyboardAvoidingView
+      h={{
+        base: '100%',
+        lg: 'auto',
+      }}
+      behavior="padding"
+    >
+      <ScrollView flex={1} p={1} keyboardShouldPersistTaps="handled">
+        <TransactionForm
+          payload={payload}
+          loading={loading}
+          accounts={accounts}
+          goToTransactions={goToTransactions}
+          getAutocompleteAccounts={getAutocompleteAccounts}
+          loadingAutocomplete={loadingAutocomplete}
+          submit={submit}
+        />
+      </ScrollView>
+    </KeyboardAvoidingView>
+  </>
+);
+
+export default Edit;
